Refactor: merged manage_channel and manage_user into user_channel_relationship to avoid confusion with channel_management
